Fix changed translations not updating
If the user goes to the preference page and saves after a translation update, then the idea is for them to have the new translation show up (which makes sense). With the old way of only updating with a changed language code, they would need to explicitly toggle the language and back. Using `strcmp` just adds unneeded complexity, so just always assuming the translation changed when preferences are updated just simplifies things. No edge cases!